emlibc: the self-host loop closes INCLUDING floating point
EmbCC gained FP codegen (user), so math folds into the on-OS self-host set. The
OS now compiles emlibc FP math with EmbCC -- the glue math.c AND all 31 lifted
fdlibm files (real ~1-ulp library) plus crt0 + libc + app -- links them with
EmbLD into a native EMBX, and runs it. Every object, FP included, produced by
the owned toolchain on the metal.

test emlibc math selfhost -- VERIFIED on the OS: embcc built 39/39 FP units,
embld linked them (236688-byte EMBX), the OS-built binary ran exit 42 passing
real fdlibm at 1e-12 relative (incl. sin(1e8), the exact-rem_pio2 case).
test mathself -- the fast de-risk: an EMBX whose math is EmbCC-compiled (rest
gcc), reproducible via a Makefile rule.

emlibc-side to fit EmbCC (the compiler is correct; these are genuine choices):
  - fdlibm.h: expm1/finite prototypes (fdlibm relies on C89 implicit decls;
    EmbCC correctly rejects them, math.c defines finite()).
  - stdlib.c: emlibc_stdio_flush_all is WEAK + null-guarded -- a program that
    never uses stdio should not link it (a real decoupling).
  - stdio.c: %f integer/fraction split via signed long long, not unsigned.
  - kernel/process/spawn.h: SPAWN_ARGV_MAX 32 -> 64 (a 39-object link is one
    embld argv; the child 4 KiB argv page is still the runtime bound).

Two NEW EmbCC front-end gaps surfaced (sidestepped above, not blocking):
weak UNDEFINED references (embcc emits U where gcc emits w), and unsigned-64
<-> double conversion (SSE2 has no unsigned convert). Flagged for EmbCC.

+++ user/emlibc/stdio/stdio.c
@@ -225,11 +225,13 @@ static int dec_exp10(double v)
 static int build_f(char *out, double v, int prec)
 {
     double scale = 1.0; for (int i = 0; i < prec; i++) scale *= 10.0;
-    unsigned long long ip = (unsigned long long)v;
+    /* signed long long, not unsigned: SSE2 has no unsigned-64<->double convert
+     * (EmbCC rejects it), and every %f integer/fraction part fits in 63 bits. */
+    long long ip = (long long)v;
     double frac = v - (double)ip;
-    unsigned long long fp = (unsigned long long)(frac * scale + 0.5);
-    if (fp >= (unsigned long long)scale) { fp -= (unsigned long long)scale; ip++; }
-    int n = f_uint(out, ip);
+    long long fp = (long long)(frac * scale + 0.5);
+    if (fp >= (long long)scale) { fp -= (long long)scale; ip++; }
+    int n = f_uint(out, (unsigned long long)ip);
     if (prec > 0) {
         out[n++] = '.';
         char t[24]; int fl = f_uint(t, fp);
